Move function to util to avoid circular dependency
