+++ powergenome/load_profiles.py
@@ -357,27 +357,3 @@ def calc_dg_frac_load_method(dg_profile, dg_requirement, regional_load, settings
     hourly_gen = dg_profile * dg_capacity
 
     return hourly_gen
-
-
-def remove_feb_29(df: pd.DataFrame) -> pd.DataFrame:
-    """Remove Feb 29 from a wide format leap-year dataseries
-
-    Parameters
-    ----------
-    df : pd.DataFrame
-        A wide format dataframe with 8784 columns
-
-    Returns
-    -------
-    pd.DataFrame
-        The same dataframe but without the 24 hours in Feb 29 and only 8760 rows.
-    """    
-    idx_start = df.index.min()
-    idx_name = df.index.name
-    df["datetime"] = pd.date_range(start="2012-01-01", freq="H", periods=8784)
-
-    df = df.loc[~((df.datetime.dt.month == 2) & (df.datetime.dt.day == 29)), :]
-    df.index = range(idx_start, idx_start + 8760)
-    df.index.name = idx_name
-
-    return df.drop(columns=["datetime"])
